Moved useWindowSize hook into correct dir
